backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
APScheduler==3.10.4
httpx==0.27.2
jinja2==3.1.4
email-validator==2.2.0
beautifulsoup4==4.12.3